move application workload config implementatuion to k8s module
